Refactored nodejs lambda so it is easier to unit test
