+++ amplify/backend/function/submissionsLambda/lib/controllers/__mocks__/aws-sdk.ts
@@ -1,0 +1,11 @@
+export const awsSdkPromiseResponse = jest.fn().mockReturnValue(Promise.resolve(true));
+
+const putFn = jest.fn().mockImplementation(() => ({ promise: awsSdkPromiseResponse }));
+
+class DocumentClient {
+  put = putFn;
+}
+
+export const DynamoDB = {
+  DocumentClient,
+};
