+++ amplify/backend/function/submissionsLambda/src/controllers/__tests__/SubmissionsController.test.js
@@ -1,8 +1,10 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
 const tslib_1 = require("tslib");
-const app_1 = require("../app");
-const node_mocks_http_1 = (0, tslib_1.__importDefault)(require("../../src/node_modules/node-mocks-http"));
+const SubmissionsController_1 = require("../SubmissionsController");
+const node_mocks_http_1 = (0, tslib_1.__importDefault)(require("../../../src/node_modules/node-mocks-http"));
+const aws_sdk_1 = require("../__mocks__/aws-sdk");
+const db = new aws_sdk_1.DynamoDB.DocumentClient();
 describe('Express app', () => {
     describe('Post feedback', () => {
         it('Should return 400 error when vote is not boolean', () => {
@@ -10,15 +12,15 @@ describe('Express app', () => {
                 method: 'POST',
                 url: '/feedback',
                 headers: {
-                    "content-type": "application/json"
+                    'content-type': 'application/json'
                 },
                 body: {
                     vote: 1,
-                    page_path: "path"
+                    page_path: 'path'
                 }
             });
             const response = node_mocks_http_1.default.createResponse();
-            (0, app_1.postCallback)(request, response);
+            (0, SubmissionsController_1.postCallback)(request, response);
             expect(response.statusCode).toBe(400);
         });
         it('Should return 400 error when vote is not in body', () => {
@@ -26,14 +28,14 @@ describe('Express app', () => {
                 method: 'POST',
                 url: '/feedback',
                 headers: {
-                    "content-type": "application/json"
+                    'content-type': 'application/json'
                 },
                 body: {
-                    page_path: "path"
+                    page_path: 'path'
                 }
             });
             const response = node_mocks_http_1.default.createResponse();
-            (0, app_1.postCallback)(request, response);
+            (0, SubmissionsController_1.postCallback)(request, response);
             expect(response.statusCode).toBe(400);
         });
         it('Should return 400 error when path is not a string', () => {
@@ -41,7 +43,7 @@ describe('Express app', () => {
                 method: 'POST',
                 url: '/feedback',
                 headers: {
-                    "content-type": "application/json"
+                    'content-type': 'application/json'
                 },
                 body: {
                     vote: true,
@@ -49,7 +51,7 @@ describe('Express app', () => {
                 }
             });
             const response = node_mocks_http_1.default.createResponse();
-            (0, app_1.postCallback)(request, response);
+            (0, SubmissionsController_1.postCallback)(request, response);
             expect(response.statusCode).toBe(400);
         });
         it('Should return 400 error when path is not in body', () => {
@@ -57,15 +59,32 @@ describe('Express app', () => {
                 method: 'POST',
                 url: '/feedback',
                 headers: {
-                    "content-type": "application/json"
+                    'content-type': 'application/json'
                 },
                 body: {
                     vote: true
                 }
             });
             const response = node_mocks_http_1.default.createResponse();
-            (0, app_1.postCallback)(request, response);
+            (0, SubmissionsController_1.postCallback)(request, response);
             expect(response.statusCode).toBe(400);
+        });
+        it('Should call dynamodb put', async () => {
+            const request = node_mocks_http_1.default.createRequest({
+                method: 'POST',
+                url: '/feedback',
+                headers: {
+                    'content-type': 'application/json'
+                },
+                body: {
+                    vote: true,
+                    page_path: '/path/to/page'
+                }
+            });
+            const response = node_mocks_http_1.default.createResponse();
+            await (0, SubmissionsController_1.postCallback)(request, response);
+            expect(response.statusCode).toBe(200);
+            expect(db.put).toHaveBeenCalled();
         });
     });
 });
